requirements.txt
# requirements.txt

absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
Flask==3.0.2
Flask-Cors==4.0.0
Flask-Uploads==0.2.1
flatbuffers==24.3.7
grpcio==1.62.1
gunicorn==21.2.0
h5py==3.10.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
libclang==16.0.6
Markdown==3.6
MarkupSafe==2.1.5
ml-dtypes==0.3.2
mysql-connector-python==8.3.0
numpy==1.26.4
oauthlib==3.2.2
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.1
protobuf==5.26.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.11.4
six==1.16.0
sqlx-batis==2.1.0
sqlx-exec==2.2.7
termcolor==2.4.0
threadpoolctl==3.3.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
waitress==3.0.0
Werkzeug==3.0.1
wrapt==1.16.0
